fix(vite-plugin): hot-updated module that is not .vine.ts trigger Vine updates
