+++ packages/compiler/src/transform.ts
@@ -558,7 +558,7 @@ export function transformFile(
     ms.appendRight(
       ms.length(),
       /* js */`export const _rerender_only = ${vineFileCtx.renderOnly}
-export const _rerender_vcf_fn_name = "${vineFileCtx.hmrCompFnsName ?? ""}"
+export const _rerender_vcf_fn_name = "${vineFileCtx.hmrCompFnsName ?? ''}"
 import.meta.hot?.accept((mod) => {
 if (!mod){return;}
 const { _rerender_only, _rerender_vcf_fn_name } = mod;
